Move world generator registering to WorldManager
